% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cch_heartbeat.R
\name{cch_heartbeat}
\alias{cch_heartbeat}
\title{Get heartbeat}
\usage{
cch_heartbeat(...)
}
\arguments{
\item{...}{Curl options passed to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\value{
list of routes
}
\description{
Get heartbeat
}
\examples{
\dontrun{
cch_heartbeat()
}
}
