% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cch_pkgs.R
\name{cch_pkgs}
\alias{cch_pkgs}
\title{Get package based checks}
\usage{
cch_pkgs(x = NULL, limit = 10, offset = 0, ...)
}
\arguments{
\item{x}{package name, optional, if you pass in more than one
we'll do async}

\item{limit}{number of records to return. Default: 10}

\item{offset}{record number to start at. Default: 0}

\item{...}{Curl options passed to \code{\link[crul:HttpClient]{crul::HttpClient()}} or
\code{\link[crul:Async]{crul::Async()}}}
}
\value{
list of info about a package(s)
}
\description{
Get package based checks
}
\examples{
\dontrun{
x <- cch_pkgs()
x$data
x$data$summary
x$data$checks
x$data$date_updated
x$data$package
x$data$url

cch_pkgs("geojsonio")
cch_pkgs(c("geojsonio", "leaflet", "MASS"))
}
}
