% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cch_pkgs_history.R
\name{cch_pkgs_history}
\alias{cch_pkgs_history}
\title{Get historical check data for packages}
\usage{
cch_pkgs_history(x, limit = 30, offset = 0, ...)
}
\arguments{
\item{x}{package name, required, if you pass in more than one
we'll do async}

\item{limit}{number of records to return. Default: 10}

\item{offset}{record number to start at. Default: 0}

\item{...}{Curl options passed to \code{\link[crul:HttpClient]{crul::HttpClient()}} or
\code{\link[crul:Async]{crul::Async()}}}
}
\value{
list of info about a package(s)
}
\description{
Get historical check data for packages
}
\note{
this function gets historical data; for current day check data only
see \code{\link[=cch_pkgs]{cch_pkgs()}}
}
\examples{
\dontrun{
x <- cch_pkgs_history(x = "geojsonio")
x
x$data
x$data$package
x$data$history
x$data$history$summary
x$data$history$summary$any
x$data$history$check_details

# many packages
res <- cch_pkgs_history(c("geojsonio", "leaflet", "MASS"))
res

# pagination
cch_pkgs_history(x = "geojsonio", limit = 3)
cch_pkgs_history(x = "geojsonio", limit = 3, offset = 4)
}
}
