% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cch_pkgs_search.R
\name{cch_pkgs_search}
\alias{cch_pkgs_search}
\title{Search historical package check data}
\usage{
cch_pkgs_search(
  q,
  package = NULL,
  one_each = NULL,
  fields = NULL,
  limit = 30,
  offset = 0,
  ...
)
}
\arguments{
\item{q}{(character) full text query string}

\item{package}{(character) a package name. limit results to a single
package, e.g, \code{package="taxize"}}

\item{one_each}{(logical) if \code{TRUE}, return a single result for each package;
useful if you want to find out what packages match a particular query, and
don't care which day that match happened. default: \code{FALSE}}

\item{fields}{(character) vector of fields to return, e.g.,
\code{fields=c("package", "check_details")}}

\item{limit}{number of records to return. Default: 10}

\item{offset}{record number to start at. Default: 0}

\item{...}{Curl options passed to \code{\link[crul:HttpClient]{crul::HttpClient()}} or
\code{\link[crul:Async]{crul::Async()}}}
}
\value{
list of info about a package(s)
}
\description{
Search historical package check data
}
\examples{
\dontrun{
x <- cch_pkgs_search(q = "memory")
x
x$data
x$data$package
x$data
x$data$summary
x$data$summary$any
x$data$check_details

# restrict returned fields
res <- cch_pkgs_search("memory", fields = c("package", "check_details"))
res$data$check_details$output[1]
grepl('memory', res$data$check_details$output[1])

# one each, one record per package
res <- cch_pkgs_search("memory", one_each = TRUE,
  fields = c("package", "date_updated"))
res

# pagination
cch_pkgs_search("memory", limit = 3)
cch_pkgs_search("memory", limit = 3, offset = 4)
}
}
