% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchecks-package.R
\docType{package}
\name{cchecks-package}
\alias{cchecks-package}
\alias{cchecks}
\title{cchecks}
\description{
Client for the cranchecks.info API
}
\section{Docs}{

See \url{https://cranchecks.info/docs}
}

\section{Historical data}{

There's an important shortcoming of historical data. The links in the
historical data in the \code{checks} field are not date specific. If you go to a
link in historical data, for example for April 2nd, 2020, links in that set
of data link to whatever the current check data is for that package. The
\code{check_details} field is date specific though; the text is scraped from the
package checks page each day and stored, so you can count on that to be date
specific. There are sometimes links to further checks, often of compiled
packages on various types of checks that CRAN runs; we do not have those
check results - we could get them but have not take the time to sort that
out.
}

\author{
Scott Chamberlain \email{sckott@protonmail.com}
}
\keyword{package}
