% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchn_register.R
\name{cchn_register}
\alias{cchn_register}
\title{Notifications: register your email address and get a token}
\usage{
cchn_register(email = NULL, token = NULL, ...)
}
\arguments{
\item{email}{(character) email address to use for interaction with
the CRAN checks API. If no email address is given, we go through a few
steps: check for the cached file mentioned below for any emails; use
\code{\link[whoami:email_address]{whoami::email_address()}}; if the location you are running this in is
a package, we look for the maintainer's email address in the package.}

\item{token}{(character) your CRAN checks API token. you shouldn't need
to pass a token here. if you used \code{\link[=cchn_register]{cchn_register()}} your token should
be cached}

\item{...}{Curl options passed to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
\code{NULL} - nothing returned
}
\description{
Notifications: register your email address and get a token
}
\details{
We cache a file with email addresses and tokens at the path
\code{file.path(rappdirs::user_data_dir("cranchecks", "cchecks"), "emails.csv")}
You can run that in R to get the path for the file on your machine.

To get a new token for an email address that was previously registered,
go to the file above and delete the line with the email address and token
for the email address in question; remember to save the change.
Then when you run \code{cchn_register()} again for that email you can get
a new token.

To add an email address that was validated before (probably on
another machine), to the configuration file, call this function
with the ‘email’ and ‘token’ arguments.
}
