% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchn_rules.R
\name{cchn_rules}
\alias{cchn_rules}
\alias{cchn_rule_add}
\alias{cchn_rule_list}
\alias{cchn_rule_get}
\alias{cchn_rule_delete}
\title{Notifications: add, list, get, delete notification rules}
\usage{
cchn_rule_add(
  status = NULL,
  platform = NULL,
  time = NULL,
  regex = NULL,
  package = NULL,
  email = NULL,
  path = ".",
  ...
)

cchn_rule_list(email = NULL, path = ".", ...)

cchn_rule_get(id, email = NULL, path = ".", ...)

cchn_rule_delete(id, email = NULL, path = ".", ...)
}
\arguments{
\item{status}{(character) a check status, one of: error, warn, note, fail}

\item{platform}{(character) a platform, a string to match against the
platform strings used by cran checks. e.g., "osx" would match any osx
platform check results, whereas you could limit the rule to just a single
specific platform by using the target platforms exact string
"r-oldrel-osx-x86_64"}

\item{time}{(character) number of days}

\item{regex}{(character) a regex string}

\item{package}{(character) a package name. if \code{NULL}, we attempt to
get the package name from the working directory, and fail out if there's
not a valid package structure/package name}

\item{email}{(character) email address to use for interaction with
the CRAN checks API. we use the email address in the maintainers slot
of the DESCRIPTION file of your working directory. you can supply an
email address instead}

\item{path}{(character) path to a directory containing an R package}

\item{...}{Curl options passed to \link[crul:verb-GET]{crul::verb-GET}, \link[crul:verb-POST]{crul::verb-POST}, or
\link[crul:verb-DELETE]{crul::verb-DELETE}}

\item{id}{(integer) a rule id. note that you can not get or delete
rules that are not yours. required}
}
\value{
\itemize{
\item \code{cchn_rule_add()}: \code{list(error = NULL, data = "success")}
\item \code{cchn_rule_get()}: list with elements \code{error} and \code{data} (a list of
the parts of the rule)
\item \code{cchn_rule_list()}: list with elements \code{error} and \code{data} (a data.frame
of all the rules associated with your email)
\item \code{cchn_rule_delete()}: if deletion works, a message saying "ok"
}
}
\description{
Notifications: add, list, get, delete notification rules
}
\details{
\itemize{
\item \code{cchn_rule_add()}: add a rule, one rule per function call
\item \code{cchn_rule_get()}: get a rule by rule id (see \code{cchn_rule_list} to get ids;
can only get rules for the authenticated user)
\item \code{cchn_rule_list()}: list all rules for the authenticated user
\item \code{cchn_rule_delete()}: delete a rule by rule id (only those for the
authenticated user)
}
}
\section{example rules}{

\itemize{
\item ERROR for 3 days in a row across 2 or more platforms
\itemize{
\item \code{cchn_rule_add(status = 'error', time = 3, platform = 2)}
}
\item ERROR for 2 days in a row on all osx platforms
\itemize{
\item \verb{cchn_rule_add(status' = 'error', time = 2, platform = "osx")}
}
\item ERROR for 2 days in a row on all release R versions
\itemize{
\item \code{cchn_rule_add(status = 'error', time = 2, platform = "release")}
}
\item WARN for 4 days in a row on any platform except Solaris
\itemize{
\item \code{cchn_rule_add(status = 'warn', time = 4, platform = "-solaris")}
}
\item WARN for 2 days in a row across 9 or more platforms
\itemize{
\item \code{cchn_rule_add(status = 'warn', time = 2, platform = 10)}
}
\item NOTE across all osx platforms
\itemize{
\item \code{cchn_rule_add(status = 'note', platform = "osx")}
}
\item NOTE
\itemize{
\item \code{cchn_rule_add(status = 'note')}
}
\item error details contain regex 'install'
\itemize{
\item \code{cchn_rule_add(regex = "install")}
}
}
}

\examples{
\dontrun{
# (x <- cchn_rule_list())
# cchn_rule_get(x$data$id[1])
# cchn_rule_delete(x$data$id[1])
# cchn_rule_get(id = x$data$id[1])
}
}
