% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchecks-package.R
\docType{package}
\name{cchecks-package}
\alias{cchecks-package}
\alias{cchecks}
\title{cchecks - Client for the cranchecks.info API}
\section{Docs}{

See \url{https://cranchecks.info/docs}
}

\author{
Scott Chamberlain \email{sckott@protonmail.com}
}
\keyword{package}
