% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchn_register.R
\name{cchn_register}
\alias{cchn_register}
\title{Notifications: register your email address and get a token}
\usage{
cchn_register(email, ...)
}
\arguments{
\item{email}{(character) email address to use for interaction with
the CRAN checks API}

\item{...}{Curl options passed to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
list, with \code{email} and \code{token} slots
}
\description{
Notifications: register your email address and get a token
}
\details{
After running this function successfully (no errors), save your
token in your .Renviron, .bash_profile, or similar.

We cache a file with details at the path
\code{file.path(rappdirs::user_data_dir("cranchecks", "cchecks"), "emails.csv")}
}
