% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchn_rules.R
\name{cchn_rules}
\alias{cchn_rules}
\alias{cchn_rule_add}
\alias{cchn_rule_list}
\alias{cchn_rule_get}
\alias{cchn_rule_delete}
\title{Notifications: add, list, get, delete notification rules}
\usage{
cchn_rule_add(
  package,
  status = NULL,
  platform = NULL,
  time = NULL,
  regex = NULL,
  token = NULL,
  ...
)

cchn_rule_list(token = NULL, ...)

cchn_rule_get(id, token = NULL, ...)

cchn_rule_delete(id, token = NULL, ...)
}
\arguments{
\item{package}{(character) a package name. required}

\item{status}{(character) a check status, one of: error, warn, note, fail}

\item{platform}{(character) a platform, a string to match against the
platform strings used by cran checks. e.g., "osx" would match any osx
platform check results, whereas you could limit the rule to just a single
specific platform by using the target platforms exact string
"r-oldrel-osx-x86_64"}

\item{time}{(character) number of days}

\item{regex}{(character) a regex string}

\item{token}{(character) your CRAN checks API token. ideally you do not
pass in your token as a parameter, but rather store as a env var}

\item{...}{Curl options passed to \link[crul:verb-GET]{crul::verb-GET}}

\item{id}{(integer) a rule id. note that you can not get or delete
rules that are not yours. required}
}
\value{
list
}
\description{
Notifications: add, list, get, delete notification rules
}
\details{
\itemize{
\item \code{cchn_rule_add()}: add a rule, one rule per function call
\item \code{cchn_rule_get()}: get a rule by rule id (see \code{cchn_rule_list} to get ids;
can only get rules for the authenticated user)
\item \code{cchn_rule_list()}: list all rules for the authenticated user
\item \code{cchn_rule_delete()}: delete a rule by rule id (only those for the
authenticated user)
}
}
\examples{
\dontrun{
# (x <- cchn_rule_list())
# cchn_rule_get(x$data$id[1])
# cchn_rule_delete(x$data$id[1])
# cchn_rule_get(id = x$data$id[1])
}
}
